requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.23.0
SQLAlchemy==1.4.52
psycopg2-binary==2.9.7
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.11
